refactor(auth): Refactor log in/out controllers - Moves both login and logout controllers to their own file
