+++ api/server/controllers/auth.controller.js
@@ -1,52 +1,10 @@
 const {
-  loginUser,
-  logoutUser,
   registerUser,
   requestPasswordReset,
   resetPassword
 } = require('../services/auth.service');
 
 const isProduction = process.env.NODE_ENV === 'production';
-
-const loginController = async (req, res) => {
-  try {
-    const token = req.user.generateToken();
-    const user = await loginUser(req.user);
-    if (user) {
-      res.cookie('token', token, {
-        expires: new Date(Date.now() + eval(process.env.SESSION_EXPIRY)),
-        httpOnly: false,
-        secure: isProduction
-      });
-      res.status(200).send({ token, user });
-    } else {
-      return res.status(400).json({ message: 'Invalid credentials' });
-    }
-  } catch (err) {
-    console.log(err);
-    return res.status(500).json({ message: err.message });
-  }
-};
-
-const logoutController = async (req, res) => {
-  const { signedCookies = {} } = req;
-  const { refreshToken } = signedCookies;
-  try {
-    const logout = await logoutUser(req.user, refreshToken);
-    console.log(logout);
-    const { status, message } = logout;
-    if (status === 200) {
-      res.clearCookie('token');
-      res.clearCookie('refreshToken');
-      res.status(status).send({ message });
-    } else {
-      res.status(status).send({ message });
-    }
-  } catch (err) {
-    console.log(err);
-    return res.status(500).json({ message: err.message });
-  }
-};
 
 const registrationController = async (req, res) => {
   try {
@@ -107,10 +65,9 @@ const resetPasswordController = async (req, res) => {
   }
 };
 
-const refreshController = async () => {
-  // const refreshController = async (req, res, next) => {
-  // const { signedCookies = {} } = req;
-  // const { refreshToken } = signedCookies;
+const refreshController = async (req, res, next) => {
+  const { signedCookies = {} } = req;
+  const { refreshToken } = signedCookies;
   //TODO
   // if (refreshToken) {
   //   try {
@@ -121,6 +78,7 @@ const refreshController = async () => {
   //         if (user) {
   //           // Find the refresh token against the user record in database
   //           const tokenIndex = user.refreshToken.findIndex(item => item.refreshToken === refreshToken);
+
   //           if (tokenIndex === -1) {
   //             res.statusCode = 401;
   //             res.send('Unauthorized');
@@ -159,8 +117,6 @@ const refreshController = async () => {
 
 module.exports = {
   getUserController,
-  loginController,
-  logoutController,
   refreshController,
   registrationController,
   resetPasswordRequestController,
